Refactor App.js to use two custom hooks
App.js has been remade to incorporate two new custom hooks - useCurrencyConverter for calculating the exchange rate and useFetch to fetch the currency data
from exchangerates api. This should make the hooks less heavy to use and more specific in their intent, increasing modularity of the app and reusability of the components.